@@ -1,17 +1,17 @@
 {
   "nativescript": {
     "id": "org.nativescript.demo",
-    "tns-ios": {
-      "version": "2.1.1"
-    },
     "tns-android": {
-      "version": "2.1.1"
+      "version": "3.0.0"
+    },
+    "tns-ios": {
+      "version": "3.0.1"
     }
   },
   "dependencies": {
     "nativescript-color-picker": "^1.3.0",
-    "nativescript-drawingpad": "file:..",
-    "tns-core-modules": "^2.0.0"
+    "nativescript-drawingpad": "file:///Users/peter/Documents/NativeScriptProjects/nativescript-drawingpad",
+    "tns-core-modules": "^3.0.0"
   },
   "devDependencies": {
     "babel-traverse": "6.7.6",
@@ -19,7 +19,7 @@
     "babylon": "6.7.0",
     "filewalker": "0.1.2",
     "lazy": "1.0.11",
-    "nativescript-dev-typescript": "^0.3.2",
-    "typescript": "^1.8.10"
+    "nativescript-dev-typescript": "^0.4.5",
+    "typescript": "^2.3.2"
   }
 }
