@@ -1,13 +1,19 @@
 {
   "name": "nativescript-drawingpad",
-  "version": "1.1.2",
+  "version": "2.0.0",
   "main": "drawingpad",
   "typings": "index.d.ts",
   "description": "A NativeScript plugin to provide a way to capture any drawing (signatures are a common use case) from the device screen.",
   "nativescript": {
     "platforms": {
-      "android": "2.1.1",
-      "ios": "2.1.1"
+      "android": "3.0.0",
+      "ios": "3.0.0"
+    },
+    "plugin": {
+      "nan": "true",
+      "pan": "true",
+      "core3": "true",
+      "category": "Interface"
     }
   },
   "repository": {
@@ -39,12 +45,17 @@
       "name": "Nathan Walker",
       "email": "walkerrunpdx@gmail.com",
       "url": "https://github.com/NathanWalker"
+    },
+    {
+      "name": "Peter Staev",
+      "email": "peter@tangrasoft.com",
+      "url": "https://github.com/PeterStaev"
     }
   ],
   "scripts": {
     "build": "tsc",
     "clean": "rm -rf demo/node_modules demo/lib demo/platforms",
-    "demo.ios": "npm run preparedemo; cd demo; tns emulate ios",
+    "demo.ios": "npm run preparedemo; cd demo; tns run ios",
     "demo.android": "npm run preparedemo; cd demo; tns run android",
     "preparedemo": "npm run build; cd demo; tns plugin remove nativescript-drawingpad; tns plugin add ..; tns install",
     "setup": "npm i; cd demo; npm install; cd ..; npm run build; cd demo; tns plugin add ..; cd .."
@@ -54,7 +65,12 @@
     "url": "https://github.com/bradmartin/nativescript-drawingpad/issues"
   },
   "homepage": "https://github.com/bradmartin/nativescript-drawingpad",
+  "peerDependencies": {
+    "tns-core-modules": "^3.0.0"
+  },
   "devDependencies": {
-    "typescript": "^1.8.10"
+    "typescript": "~2.2.0",
+    "tns-core-modules": "3.0.0",
+    "tns-platform-declarations": "3.0.1"
   }
 }
