@@ -1,108 +1,80 @@
-﻿import { PropertyChangeData } from "ui/core/dependency-observable";
-import { PropertyMetadata } from "ui/core/proxy";
-import { ContentView } from "ui/content-view";
-import { Color } from "color";
-import common = require("./drawingpad-common");
+﻿import { Color } from "color";
+import { DrawingPadBase, penColorProperty, penWidthProperty } from "./drawingpad-common";
 
-declare var SignatureView: any, CGRectMake: any;
+declare var SignatureView: any;
 
-export class DrawingPad extends ContentView {
-  private _ios: any = SignatureView;
-  private _penColor: string;
-  private _penWidth: number;
+export * from "./drawingpad-common";
 
-  constructor() {
-    super();
-    // console.log('--------- DrawingPad ---------');
-    this._ios = SignatureView.alloc().initWithFrame(CGRectMake(0, 0, 100, 100));
-    this._ios.clipsToBounds = true;
-  }
-
-  get ios(): any {
-    return this._ios;
-  }
-
-  get _nativeView(): any {
-    return this._ios;
-  }
-
-  set penColor(value: string) {
-    if (this._ios && value) {
-      // console.log(`set penColor: ${value}`);
-      this._ios.setLineColor(new Color(value).ios);
-    } else {
-      this._penColor = value;
-    }
-  }
-
-  set penWidth(value: any) {
-    if (this._ios) {
-      if (value && typeof value !== 'undefined' && value !== NaN && value !== 'NaN') {
-        // console.log(`set penWidth: ${value}`);
-        this._ios.setLineWidth(Math.floor(parseInt(value)));
-      }
-    } else {
-      this._penWidth = value;
-    }
-  }
-
-  public onLoaded() {
-
-    // console.log(`onLoaded ${this.width}, ${this.height}`);
-    if (this.width) {
-      this._ios.frame.size.width = this.width;
-    }
-    if (this.height) {
-      this._ios.frame.size.height = this.height;
+export class DrawingPad extends DrawingPadBase {
+    constructor() {
+        super();
+        // console.log('--------- DrawingPad ---------');
+        this.nativeView = SignatureView.alloc().initWithFrame(CGRectMake(0, 0, 100, 100));
+        this.nativeView.clipsToBounds = true;
     }
 
-    try {
-      if (this._penColor) {
-        this.penColor = this._penColor;
-      }
-
-      if (this._penWidth) {
-        this.penWidth = this._penWidth;
-      }
-
-    } catch (ex) {
-      console.log(ex);
+    get ios(): any {
+        return this.nativeView;
     }
 
-  }
+    [penWidthProperty.getDefault](): number {
+        return this.nativeView.foregroundLineWidth;
+    }
+    [penWidthProperty.setNative](value: number) {
+        this.nativeView.setLineWidth(Math.floor(value));
+    }
 
-  public getDrawing(): Promise<any> {
-    return new Promise((resolve, reject) => {
-      try {
-        let isSigned = this._ios.isSigned();
-        if (isSigned === true) {
-          let data = this._ios.signatureImage();
-          resolve(data);
-        } else {
-          reject("DrawingPad is empty.");
+    [penColorProperty.getDefault](): UIColor {
+        return this.nativeView.foregroundLineColor;
+    }
+    [penColorProperty.setNative](value: Color | UIColor) {
+        const color = value instanceof Color ? value.ios : value;
+        this.nativeView.setLineColor(color);
+    }
+
+    public onLoaded() {
+
+        // console.log(`onLoaded ${this.width}, ${this.height}`);
+        if (this.width) {
+            this.nativeView.frame.size.width = this.width;
         }
-      } catch (err) {
-        reject(err);
-      }
-    })
-  }
-
-  public clearDrawing(): any {
-    try {
-      if (this.backgroundColor) {
-        let color = this.backgroundColor;
-        if (color.constructor == Color) {
-          color = color.ios;
-        } else if (color.constructor == String) {
-          color = new Color(<any>color).ios;
-        } 
-        this._ios.clearWithColor(color);
-      } else {
-        this._ios.clear();
-      }
-    } catch (err) {
-      console.log("Error clearing the DrawingPad: " + err);
+        if (this.height) {
+            this.nativeView.frame.size.height = this.height;
+        }
     }
-  }
+
+    public getDrawing(): Promise<any> {
+        return new Promise((resolve, reject) => {
+            try {
+                let isSigned = this.nativeView.isSigned();
+                if (isSigned === true) {
+                    let data = this.nativeView.signatureImage();
+                    resolve(data);
+                } else {
+                    reject("DrawingPad is empty.");
+                }
+            } catch (err) {
+                reject(err);
+            }
+        })
+    }
+
+    public clearDrawing(): void {
+        try {
+            if (this.backgroundColor) {
+                let color = this.backgroundColor;
+                if (color.constructor == Color) {
+                    color = color.ios;
+                } else if (color.constructor == String) {
+                    color = new Color(<any>color).ios;
+                }
+                this.nativeView.clearWithColor(color);
+            } else {
+                this.nativeView.clear();
+            }
+        } catch (err) {
+            console.log("Error clearing the DrawingPad: " + err);
+        }
+    }
 
 }
